backend/requirements.txt
alembic
python-dotenv
psycopg2
fastapi>=0.100.0
pydantic>=2.4.0
email-validator
passlib[bcrypt]
SQLAlchemy>=2.0.0,<3.0.0
uvicorn[standard]
python-jose[cryptography]
python-multipart
pyotp
aiokafka
bcrypt>=4.0.1
cffi>=1.16.0
wheel
